requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.5.0
chromadb==0.4.18
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
httpx==0.25.2 